requirements.txt
ansicon==1.89.0
APScheduler==3.11.0
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
blessed==1.20.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
cssselect2==0.7.0
defusedxml==0.7.1
Django==5.1.3
django-background-tasks==1.2.8
django-picklefield==3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
fonttools==4.55.1
fpdf2==2.8.1
google-auth==2.36.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
html5lib==1.1
httplib2==0.22.0
idna==3.10
jinxed==1.3.0
lxml==5.3.0
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
oscrypto==1.3.0
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.1
pyparsing==3.2.0
pypdf==5.1.0
pyphen==0.15.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
qrcode==8.0
redis==3.5.3
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rsa==4.9
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
stripe==11.3.0
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
wcwidth==0.2.13
weasyprint==63.0
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
